Add internal IndexLocator class
This replaces matplotlib's native class. Instead of starting ticks
from the data minimum (plus offset) we start ticks from zero. This
harmonizes IndexLocator with the other locators and prevents weird
issue where it cannot be used on empty axes. It also fixes a bug
where too many ticks are drawn when "base" is not one and uses a
default "base" of 1 and default "offset" of 0 instead of making
them required arguments.