Improve 'custom' fontset, use it mathtext.default
This supports the dummy entry 'regular' for 'custom' fontset
mathtext settings 'mathtext.it', 'mathtext.rm', etc. When detected
then the currently active font is selected rather than a global
default font. It also sets 'mathtext.default' to 'it'. Previously
this would result in math text resisting changes to the global font
but now this simply italices the current font and uses it for math.